feat: 온보딩 구현 애니메이션 추가 - 온보딩 마지막 페이지에서 빵빠레 애니메이션 동작
