+++ app/src/main/java/com/example/client/onboarding/OnBoarding1stFragment.kt
@@ -1,23 +1,26 @@
-package com.example.client
+package com.example.client.onboarding
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.client.R
 
-class OnBoarding4thFragment : Fragment() {
+
+class OnBoarding1stFragment : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
     }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_on_boarding4th, container, false)
+
+        return inflater.inflate(R.layout.fragment_on_boarding1st, container, false)
     }
