+++ app/src/main/java/com/example/client/onboarding/OnBoarding2ndFragment.kt
@@ -1,13 +1,14 @@
-package com.example.client
+package com.example.client.onboarding
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.client.R
 
 
-class OnBoarding5thFragment : Fragment() {
+class OnBoarding2ndFragment : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -18,6 +19,7 @@ class OnBoarding5thFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_on_boarding5th, container, false)
+        return inflater.inflate(R.layout.fragment_on_boarding2nd, container, false)
     }
+
 }
