+++ app/src/main/java/com/example/client/onboarding/OnBoarding3rdFragment.kt
@@ -1,10 +1,12 @@
-package com.example.client
+package com.example.client.onboarding
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.client.R
+
 class OnBoarding3rdFragment : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
