+++ app/src/main/java/com/example/client/onboarding/OnBoarding4thFragment.kt
@@ -1,25 +1,24 @@
-package com.example.client
+package com.example.client.onboarding
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.client.R
 
-
-class OnBoarding1stFragment : Fragment() {
+class OnBoarding4thFragment : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
     }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-
-        return inflater.inflate(R.layout.fragment_on_boarding1st, container, false)
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_on_boarding4th, container, false)
     }
