+++ app/src/main/java/com/example/client/onboarding/OnBoardingAdapter.kt
@@ -1,10 +1,8 @@
-package com.example.client
+package com.example.client.onboarding
 
-import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
-import androidx.fragment.app.FragmentStatePagerAdapter
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.example.client.OnBoardingActivity
 
 class OnBoardingAdapter(onBoardingActivity: OnBoardingActivity) : FragmentStateAdapter(onBoardingActivity) {
