+++ app/src/main/res/anim/scale_animation.xml
@@ -1,0 +1,10 @@
+<scale xmlns:android="http://schemas.android.com/apk/res/android"
+    android:duration="400"
+    android:fromXScale="0.7"
+    android:toXScale="1.3"
+    android:fromYScale="0.7"
+    android:toYScale="1.3"
+    android:pivotX="50%"
+    android:pivotY="50%"
+    android:repeatCount="1"
+    android:repeatMode="reverse" />
