+++ app/src/main/res/layout/fragment_on_boarding1st.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".OnBoarding1stFragment"
+    tools:context=".onboarding.OnBoarding1stFragment"
     android:background="@color/white">
 
     <LinearLayout
