+++ app/src/main/res/layout/fragment_on_boarding3rd.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".OnBoarding3rdFragment"
+    tools:context=".onboarding.OnBoarding3rdFragment"
     android:background="@color/white">
 
     <LinearLayout
