+++ app/src/main/res/layout/fragment_on_boarding5th.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".OnBoarding5thFragment"
+    tools:context=".onboarding.OnBoarding5thFragment"
     android:background="@color/white">
 
     <LinearLayout
@@ -63,6 +63,7 @@
         android:layout_marginTop="43dp"
         />
     <ImageView
+        android:id="@+id/iv_confetti"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:src="@drawable/img_onboarding5_confetti"
